Added support for Toshiba laptop
Added keyboard shortcuts and functionality for screen brightness,
touchpad enable/disable and the keyboard backlight for the Toshiba
Portege Z30-B-117 laptop. Files "brightness.sh" and
"toggle_keyboard_backlight.sh" need to be run with sudo and thus need to
be added to /etc/sudoers file in order for the shortcuts to work!